Sistemata la pagina di cambio password

+++ php/recoverPassword.php
@@ -35,40 +35,49 @@
                 <h2>Reimposta la password</h2>
             </div>
             <div class="regform-main-section">
-            <?php 
+            <?php
                     include_once ('User.php');
 
                     if(isset($_POST['submit'])){
                         $result = User::passwordRecoveryChange($_POST['change_pw_token'], $_POST['password']);
                         if($result) {
                             echo '<span>Password cambiata con successo!</span><br/>';
+                            echo '<ul id="regform-links">
+                                <li><a href="index.php">Tona alla Home</a></li>
+                                <li><a href="login.php">Torna al login</a></li>
+                            </ul>';
                         } else {
                             echo '<span>C\'è stato un errore nel recuperare la password, controlla che il token sia corretto o che non sia scaduto :(</span><br/>';
                         }
-                    } 
-                ?>
-                <form action="<?php echo $_SERVER['PHP_SELF']; ?>" method="POST">
-                <label for="lpassword">Password</label>
-                <input class="profile-input" type="password" id="lpassword" name="password" placeholder="Password" />
+                        die();
+                    }
+            //Controllo se il token esiste, altrimenti non stampo il form ma un messaggio di errore
+            if(
+                (User::checkIfTokenExist($_GET['token']))
+                ||
+                (User::checkIfTokenExist($_POST['change_pw_token']))
+            ){ //Il token esiste
+                echo '<form action="'.$_SERVER['PHP_SELF'].'" method="POST">
+                    <label for="lpassword">Password</label>
+                    <input class="profile-input" type="password" id="lpassword" name="password" placeholder="Password" />
 
-                <label for="lpassword-confirm">Conferma la Password</label>
-                <input class="profile-input" type="password" id="lpassword-confirm" name="password" placeholder="Conferma la Password" />
-                
-                <?php
-                if(isset($_GET['token'])){ //Alla prima apertura stampo il token dal get
-                    echo '<input type="hidden" name="change_pw_token" value="'.$_GET['token'].'" />';
-                }else{ //Nel caso di riapertura della pagina dopo la richiesta lo prendo dal post
-                    echo '<input type="hidden" name="change_pw_token" value="'.$_POST['change_pw_token'].'" />';
-                }
-                ?>
-                <input class="profile-input" name="submit" type="submit" value="Reimposta" />
-                </form>
+                    <label for="lpassword-confirm">Conferma la Password</label>
+                    <input class="profile-input" type="password" id="lpassword-confirm" name="password" placeholder="Conferma la Password" />
+                    ';
+                    if(isset($_GET['token'])){ //Alla prima apertura stampo il token dal get
+                        echo '<input type="hidden" name="change_pw_token" value="'.$_GET['token'].'" />';
+                    }else{ //Nel caso di riapertura della pagina dopo la richiesta lo prendo dal post
+                        echo '<input type="hidden" name="change_pw_token" value="'.$_POST['change_pw_token'].'" />';
+                    }
+                    
+                    echo '<input class="profile-input" name="submit" type="submit" value="Reimposta" />
+                </form>';
+            }else{ //Il token non esiste
+                echo 'Purtroppo il token fornito per il cambio della password non è valido... Controlla o richiedine uno nuovo...';
+            }
+            ?>
             </div>
-            <div class="regform-side-section">
-                <p>Non sei ancora registrato?
-                <p>Clicca <a href='registrazione.php'>qui</a> per creare un nuovo account.</p>
-            </div>
-            <ul>
+            <ul id="regform-links">
                 <li><a href="index.php">Home</a></li>
                 <li><a href="index.php">About</a></li>
             </ul>
